+++ packages/ub-pub/i18n.js
@@ -15,7 +15,47 @@ const __i18n = {
  * @returns {*}
  */
 function i18n (localeString) {
-  return __i18n[localeString] || localeString
+  let res = __i18n[localeString]
+  if (res !== undefined || localeString == null) return res
+
+  if ($App.domainInfo == null) {
+    // Domain is not loaded yet, cannot resolve string to entity or entity attribute's name
+    return localeString
+  }
+
+  // Try to resolve string as entity name or entity attribute name
+  let entity,
+      attr,
+      parts = localeString.split('.')
+
+  if (parts.length > 2) {
+    // String contain more than one dots, that is not what can be resolved to entity or entity attribute name
+    return localeString
+  }
+
+  entity = $App.domainInfo.entities[parts[0]]
+  if (!entity) {
+    // First part shall be a valid entity name
+    return localeString
+  }
+
+  if (parts.length === 1) {
+    // A valid entity name, resolve to the entity's caption
+    // Remember in __i18n for performance
+    __i18n[localeString] = entity.caption
+    return entity.caption
+  }
+
+  attr = entity.attributes[parts[1]];
+  if (!attr) {
+    // Expecting the second part to be a valid entity attribute name
+    return localeString
+  }
+
+  // A valid entity attribute name, resolve to the entity attribute's caption
+  // Remember in __i18n for performance
+  __i18n[localeString] = attr.caption
+  return attr.caption
 }
 
 /**